Add files via upload
HR Trend Analytics is a project created using Ms excel and Power Bi . The data is provided by the company . the company HR want to analyse the trend of presence and absence of employees in company. the data is formatted in Ms excel and clean and clear visual dashboard is created using power Bi software. the insights should be easily obtained by analysing the trends in dashboard.